move money to its own folder
