+++ app/lib/money/index.ts
@@ -1,0 +1,1 @@
+export { Money } from './money';
