+++ app/lib/money/money.ts
@@ -1,42 +1,13 @@
 import { Big } from 'big.js';
-
-export type NumberInput = number | string | Big;
-
-export type MoneyInput = {
-  /**
-   * Money amount
-   */
-  amount: NumberInput;
-  /**
-   * Currency for the provided amount of money
-   */
-  currency: string;
-  /**
-   * Unit of currency to use. For example for USD it can be 'usd' or 'cent', for BTC 'btc', 'sat' or 'sat', etc.
-   * If not provided the default/base unit is used (bitcoin for BTC, dollar for USD, etc.)
-   */
-  unit?: string;
-};
-
-type FormatOptions = {
-  locale?: string;
-  currency?: string;
-};
-
-type CurrencyUnit = {
-  name: string;
-  decimals: number;
-  symbol: string;
-  factor: Big;
-  format: (value: number, options?: FormatOptions) => string;
-};
-
-type CurrencyData = {
-  baseUnit: string;
-  units: CurrencyUnit[];
-};
-
-type BaseFormatOptions = FormatOptions & { decimals: number };
+import type {
+  BaseFormatOptions,
+  CurrencyData,
+  CurrencyUnit,
+  FormatOptions,
+  MoneyData,
+  MoneyInput,
+  NumberInput,
+} from './types';
 
 function baseFormat(value: number, options: BaseFormatOptions) {
   const { locale, decimals, currency } = options;
@@ -141,26 +112,6 @@ const getCurrencyBaseUnit = (currency: string) => {
     throw new Error(`Misconfigured currency: ${currency}`);
   }
   return baseUnit;
-};
-
-type MoneyData = {
-  /**
-   * Currency of the money
-   */
-  currency: string;
-  /**
-   * Amount of the money stored in the `unit` format
-   */
-  amount: Big;
-  /**
-   * Unit in which the `amount` is stored. It is always the minimal supported unit for the `currency` (e.g. millisatoshi
-   * for BTC)
-   */
-  amountUnit: CurrencyUnit;
-  /**
-   * Unit of the initial amount that was provided when creating the money
-   */
-  initialUnit: CurrencyUnit;
 };
 
 export class Money {
